# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.CreateDMESysIDRequest:
    properties:
      description:
        example: Primary production environment for marina operations
        type: string
      isActive:
        example: true
        type: boolean
      marinaId:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      name:
        example: Production System
        type: string
      organizationId:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      systemId:
        example: SYS123456
        type: string
    required:
    - name
    - organizationId
    - systemId
    type: object
...
  requests.LinkDMESysIDRequest:
    properties:
      marinaId:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
    required:
    - marinaId
    type: object
...
  requests.UpdateDMESysIDRequest:
    properties:
      description:
        example: Primary production environment for marina operations
        type: string
      isActive:
        example: true
        type: boolean
      marinaId:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      name:
        example: Production System
        type: string
      systemId:
        example: SYS123456
        type: string
    type: object
...
  responses.DMESysIDListResponseWrapper:
    properties:
      data:
        items:
          $ref: '#/definitions/responses.DMESysIDResponse'
        type: array
    type: object
  responses.DMESysIDResponse:
    description: DME System IDs used for marina integration
    properties:
      createdAt:
        type: string
      description:
        example: Primary production environment
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      isActive:
        example: true
        type: boolean
      marinaId:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      name:
        example: Production System
        type: string
      organizationId:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      systemId:
        example: SYS123456
        type: string
      updatedAt:
        type: string
    type: object
  responses.DMESysIDResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/responses.DMESysIDResponse'
    type: object
...
  responses.Error:
    description: Error response structure with optional validation details
    properties:
      details:
        items:
          $ref: '#/definitions/responses.ValidationError'
        type: array
      message:
        example: Validation failed
        type: string
    type: object
...
  responses.ValidationError:
    description: Specific validation error for a single field
    properties:
      field:
        example: username
        type: string
      message:
        example: Username is required
        type: string
      tag:
        example: required
        type: string
      value:
        example: invalid_value
        type: string
    type: object
...
info:
  contact: {}
  description: This is a API Server.
  title: DockMaster Web API
  version: 0.0.1
paths:
...
  /dme/sysids:
    get:
      consumes:
      - application/json
      description: Get a list of all DME System IDs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMESysIDListResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: List all DME System IDs
      tags:
      - DME System IDs
    post:
      consumes:
      - application/json
      description: Create a new DME System ID for an organization
      parameters:
      - description: Create DME System ID request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateDMESysIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMESysIDResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Create DME System ID
      tags:
      - DME System IDs
  /dme/sysids/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a DME System ID
      parameters:
      - description: DME System ID identifier
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Delete DME System ID
      tags:
      - DME System IDs
    get:
      consumes:
      - application/json
      description: Get DME System ID by its unique identifier
      parameters:
      - description: DME System ID identifier
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMESysIDResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Get DME System ID by ID
      tags:
      - DME System IDs
    put:
      consumes:
      - application/json
      description: Update a DME System ID
      parameters:
      - description: DME System ID identifier
        in: path
        name: id
        required: true
        type: string
      - description: Update DME System ID request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateDMESysIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMESysIDResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Update DME System ID
      tags:
      - DME System IDs
  /dme/sysids/{id}/link:
    patch:
      consumes:
      - application/json
      description: Link a DME System ID to a specific marina
      parameters:
      - description: DME System ID identifier
        in: path
        name: id
        required: true
        type: string
      - description: Link DME System ID request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.LinkDMESysIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMESysIDResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Link DME System ID to marina
      tags:
      - DME System IDs
  /dme/sysids/{id}/unlink:
    patch:
      consumes:
      - application/json
      description: Remove the link between a DME System ID and a marina
      parameters:
      - description: DME System ID identifier
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMESysIDResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Unlink DME System ID from marina
      tags:
      - DME System IDs
  /dme/sysids/marina/{marinaId}:
    get:
      consumes:
      - application/json
      description: Get DME System ID linked to a specific marina
      parameters:
      - description: Marina ID
        in: path
        name: marinaId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMESysIDResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Get DME System ID by marina ID
      tags:
      - DME System IDs
  /dme/sysids/organization/{organizationId}:
    get:
      consumes:
      - application/json
      description: Get all DME System IDs for an organization
      parameters:
      - description: Organization ID
        in: path
        name: organizationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMESysIDListResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Get DME System IDs by organization ID
      tags:
      - DME System IDs
  /dme/sysids/system/{systemId}:
    get:
      consumes:
      - application/json
      description: Get DME System ID by its external system identifier
      parameters:
      - description: System identifier
        in: path
        name: systemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMESysIDResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Get DME System ID by system identifier
      tags:
      - DME System IDs
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"